feat: better popover types
