+++ packages/popover/Popover.stories.tsx
@@ -1,10 +1,7 @@
+import { useCallback, useRef, useState } from 'react'
 import { Story, Meta } from '@storybook/react'
 import { Button } from '@lidofinance/button'
-import { PopoverProps, PopoverPlacement, PopoverOffset } from './types'
-
-import Popover from './Popover'
-import { useCallback, useRef, useState } from 'react'
-import { DEFAULT_PLACEMENT } from './constants'
+import { Popover, DEFAULT_PLACEMENT, PopoverProps, PopoverPlacement, PopoverOffset } from '.'
 
 const getOptions = (enumObject: Record<string, string | number>) =>
   Object.values(enumObject).filter((value) => typeof value === 'string')
