+++ packages/popover/PopoverRootStyles.ts
@@ -1,5 +1,6 @@
 import styled, { css } from 'styled-components'
 import { TransitionInnerProps } from '@lidofinance/transition'
+import { InferStyledComponentPropsWithoutRef } from '@lidofinance/utils'
 
 type TransitionProps = {
   $duration: number
@@ -45,6 +46,8 @@ const hiddenCSS = css`
 const getTransitionCSS = ({ $transition }: TransitionProps) => {
   return ['exiting', 'exited'].includes($transition) ? hiddenCSS : visibleCSS
 }
+
+export type PopoverRootStyleProps = InferStyledComponentPropsWithoutRef<typeof PopoverRootStyle>
 
 export const PopoverRootStyle = styled.div<TransitionProps>`
   box-sizing: border-box;
