+++ packages/popover/constants.ts
@@ -9,3 +9,31 @@ export const INITIAL_STYLE: CSSProperties = {
 }
 
 export const DEFAULT_PLACEMENT = 'bottomLeft'
+
+export enum PopoverOffset {
+  xs,
+  sm,
+  md,
+  lg,
+  xl,
+}
+export type PopoverOffsets = keyof typeof PopoverOffset
+
+export enum PopoverPlacement {
+  topLeft,
+  top,
+  topRight,
+
+  rightTop,
+  right,
+  rightBottom,
+
+  bottomLeft,
+  bottom,
+  bottomRight,
+
+  leftTop,
+  left,
+  leftBottom,
+}
+export type PopoverPlacements = keyof typeof PopoverPlacement
