+++ packages/popover/usePopoverPosition.ts
@@ -1,6 +1,6 @@
 import React, { useRef, useLayoutEffect, useState, CSSProperties } from 'react'
 import { useWindowSize } from '@lidofinance/hooks'
-import { PopoverRootProps } from './types'
+import { PopoverRootProps } from './PopoverRoot'
 import { INITIAL_STYLE, DEFAULT_PLACEMENT } from './constants'
 import { calculatePosition } from './calculatePosition'
